refactor: 이미지 엔티티 리팩토링 - 불필요한 컬럼 삭제 및 컬럼 추가 - 디렉토리 변경
